Role

+++ app/Http/Controllers/Dashboard/BloodTypeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
 use App\Models\BloodType;
+use App\Models\Client;
 use Illuminate\Http\Request;
 
 class BloodTypeController extends Controller
@@ -13,8 +14,12 @@ class BloodTypeController extends Controller
      */
     public function index()
     {
-        $bloodTypes = BloodType::all();
-        return view('bloodType.index',compact('bloodTypes'));
+        $this->authorize('viewAny', BloodType::class);
+
+        $bloodTypes = BloodType::with('donations')->get();
+        dd($bloodTypes->clients);
+        $clients = Client::where()->get();
+        return view('admin.bloodType.index',compact('bloodTypes'));
     }
 
     /**
@@ -22,7 +27,9 @@ class BloodTypeController extends Controller
      */
     public function show(BloodType $bloodType)
     {
-        return view('bloodType.show', compact('bloodType'));
+        $this->authorize('view', BloodType::class);
+
+        return view('admin.bloodType.show', compact('bloodType'));
     }
 
     /**
@@ -30,7 +37,9 @@ class BloodTypeController extends Controller
      */
     public function edit(BloodType $bloodType)
     {
-        return view('bloodType.edit', compact('bloodType'))->with('Info', 'Blood Type Updated Successfully');
+        $this->authorize('create', BloodType::class);
+
+        return view('admin.bloodType.edit', compact('bloodType'))->with('Info', 'Blood Type Updated Successfully');
     }
 
     /**
@@ -38,6 +47,8 @@ class BloodTypeController extends Controller
      */
     public function update(Request $request, BloodType $bloodType)
     {
+        $this->authorize('update', BloodType::class);
+
         $request->validate([
             'name' => 'required|string|max:255',
         ]);
@@ -50,6 +61,8 @@ class BloodTypeController extends Controller
      */
     public function destroy(BloodType $bloodType)
     {
+        $this->authorize('delete', BloodType::class);
+
         $bloodType->delete();
         return to_route('bloodType.index')->with('Danger', 'Blood Type Deleted Successfully');
     }
