+++ app/Http/Controllers/Dashboard/PostController.php
@@ -10,29 +10,25 @@ use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('permission:view-post')->only(['index', 'show']);
-        $this->middleware('permission:create-post')->only(['create', 'store']);
-        $this->middleware('permission:edit-post')->only(['edit', 'update']);
-        $this->middleware('permission:delete-post')->only('destroy');
-        $this->middleware('permission:viewAny-post')->only('index');
-
-    }
     public function index()
     {
+        $this->authorize('viewAny', Post::class);
         $posts = Post::with('category')->paginate(10);
-        return view('posts.index', compact('posts'));
+        return view('admin.posts.index', compact('posts'));
     }
 
     public function create()
     {
+        $this->authorize('create', Post::class);
+
         $categories = Category::all();
-        return view('posts.create', compact('categories'));
+        return view('admin.posts.create', compact('categories'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Post::class);
+
         $request->validate([
             'title' => 'required|string|max:255',
             'content' => 'required|string',
@@ -51,21 +47,26 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        $this->authorize('update', Post::class);
         $post->load('category');
         $categories = Category::all();
-        return view('posts.edit', compact('post', 'categories'));
+        return view('admin.posts.edit', compact('post', 'categories'));
     }
     public function show(Post $post)
     {
+        $this->authorize('view', Post::class);
+
         // $clients = Client::whereHas('posts', function ($query) use ($post) {
         //     $query->where('clientables.client_id', $post); // Corrected reference
         // })->get();
         // dd($clients);
-        return view('posts.show', compact('post'));
+        return view('admin.posts.show', compact('post'));
     }
 
     public function update(Request $request, Post $post)
     {
+        $this->authorize('update', Post::class);
+
         $request->validate([
             'title' => 'required|string|max:255',
             'content' => 'required|string',
@@ -84,6 +85,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        $this->authorize('delete', Post::class);
+
         $post->delete();
         return to_route('posts.index')->with('success', 'Post deleted successfully.');
     }
