+++ app/Http/Controllers/Dashboard/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Models\Client;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
@@ -11,18 +12,29 @@ class UserController extends Controller
 {
     public function index()
     {
-        $users = User::with('roles')->paginate();
-        return view('users.index', compact('users'));
+        $this->authorize('viewAny', user::class);
+
+        $users = user::with('roles')->paginate();
+        // $user = user::find(20);
+        // $user->getPermissionsViaRoles();
+        // foreach ($user->getPermissionsViaRoles() as $role) {
+            // dd($user->permissions);
+        // }
+        return view('admin.users.index', compact('users'));
     }
 
     public function create()
     {
+        $this->authorize('create', user::class);
+
         $roles = Role::all();
-        return view('users.create', compact('roles'));
+        return view('admin.users.create', compact('roles'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', user::class);
+
         $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users',
@@ -44,11 +56,13 @@ class UserController extends Controller
     public function edit(User $user)
     {
         $roles = Role::all();
-        return view('users.edit', compact('user', 'roles'));
+        return view('admin.users.edit', compact('user', 'roles'));
     }
 
     public function update(Request $request, User $user)
     {
+        $this->authorize('update', user::class);
+
         $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users,email,' . $user->id,
@@ -71,6 +85,8 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        $this->authorize('delete', user::class);
+
         $user->delete();
         return redirect()->route('users.index')->with('Danger', 'User deleted successfully.');
     }
